requirements.txt
torch
torchvision
opencv-python
numpy
pandas
scikit-learn
matplotlib
tqdm
pyyaml
pillow
